feat: accept default options on constructor
In `0.4.0` and below, there is a common need to pass `minecraftVersion` and `modLoader` to many subcommands. In consuming applications, this leads to prop-drilling which is tedious, verbose, and error-prone. Allowing the constructor to hold default values for these common properties eliminates the need to prop-drill. For example, this reduced the size of all affected examples by 1-8 lines or up to 14%.