feat: throw errors instead of returning `null`
Previously `undefined` and more recently `null` were returned when certain functions failed to find data. This inflicted cumbersome `null` checks in internal, example, and consumer code.

Rigorous investigation was done to consider doubling-down with Rust-style Error handling, but the result of that investigation made it obvious that throwing errors is more idiomatic in JS/TS and Deno. It is also optionally verbose rather than mandatory, and it enables destructuring in places where it was previously impossible.

To compensate for the lack of error handling in the type system which results from this decision, investigation was done to determine which errors internal libraries throw, and to document all known throwable errors on each method which could produce them, providing intellisense which should ease debugging.